refactor : auth 기능에서 순수 jpa를 spring data jpa로 리펙토링
