+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/Auth/dto/request/SignInRequestDto.java
@@ -1,17 +1,13 @@
 package com.creativedesignproject.kumoh_board_backend.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.Data;
 
-@Getter
-@Setter
-@NoArgsConstructor
+@Data
 public class SignInRequestDto {
-    @NotBlank
+    @NotBlank(message = "아이디는 필수 입력 값입니다.")
     private String userId;
 
-    @NotBlank
+    @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
     private String password;
 }
