+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/Auth/dto/request/UserIdCheckRequestDto.java
@@ -1,13 +1,9 @@
 package com.creativedesignproject.kumoh_board_backend.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.Data;
 
-@Getter
-@Setter
-@NoArgsConstructor
+@Data
 public class UserIdCheckRequestDto {
     @NotBlank(message = "아이디는 필수 입력 값입니다.")
     private String userId;
