+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/Auth/service/VerificationCodeProvider.java
@@ -1,6 +1,6 @@
 package com.creativedesignproject.kumoh_board_backend.auth.service;
 
-import com.creativedesignproject.kumoh_board_backend.auth.domain.VerificationCode;
+import com.creativedesignproject.kumoh_board_backend.auth.domain.entity.VerificationCode;
 
 public interface VerificationCodeProvider {
     VerificationCode provide();
