+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/auth/domain/entity/User.java
@@ -1,4 +1,4 @@
-package com.creativedesignproject.kumoh_board_backend.auth.domain;
+package com.creativedesignproject.kumoh_board_backend.auth.domain.entity;
 
 import com.creativedesignproject.kumoh_board_backend.common.baseentity.BaseTimeEntity;
 import com.fasterxml.jackson.annotation.JsonIgnore;
