+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/auth/domain/entity/VerificationCode.java
@@ -1,4 +1,4 @@
-package com.creativedesignproject.kumoh_board_backend.auth.domain;
+package com.creativedesignproject.kumoh_board_backend.auth.domain.entity;
 
 import java.util.regex.Pattern;
