+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/auth/domain/repository/CertificationRepository.java
@@ -1,0 +1,10 @@
+package com.creativedesignproject.kumoh_board_backend.auth.domain.repository;
+
+import com.creativedesignproject.kumoh_board_backend.auth.domain.entity.Certification;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CertificationRepository extends JpaRepository<Certification, Long> {
+    Certification findByEmail(String email);
+    void deleteByCertificationUserEmail(String email);
+}
