+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/auth/presentation/AuthController.java
@@ -1,4 +1,4 @@
-package com.creativedesignproject.kumoh_board_backend.auth.controller;
+package com.creativedesignproject.kumoh_board_backend.auth.presentation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
